refactor(conversation): unify composer autocompletes + split row renderers
Quality-review slice 4. Collapse the composer's `/` (command) and `@` (path)
autocompletes — two hand-mirrored state machines — into ONE:

- `autocomplete-machine.ts` (+test): the pure priority + Esc-dismiss-latch
  resolution; array order is priority, replacing the `!showCommands` gate.
- `use-composer-autocomplete.tsx`: a `CompletionSource` descriptor, the single
  `useComposerAutocomplete` hook, and one `CompletionPopover` replacing the twin
  `<ul>`s. Directory-drill-in and the per-token Esc latch are preserved exactly.
- `composer-sources.tsx`: the `/` and `@` sources as thin configs over the
  untouched pure cores (command-autocomplete.ts / path-autocomplete.ts).

Split the 1432-line Conversation.tsx (now 374, a thin container owning only the
turn lifecycle + transcript):

- `Composer.tsx`: draft persistence, staged images, the autocomplete hook, the
  queued-message strip, textarea + control row, onKeyDown.
- `items/`: Item dispatcher, UsageBar, message-rows (User/Assistant/Copy),
  ReasoningRow, ToolRow, WorkingRow, PermissionRow, status-rows. ToolKindIcon /
  ToolStatusGlyph switches become typed Record lookups.
- `tool-detail.ts` (+test): the branchy stringify/detail/preview helpers.

ColdThread now imports Item/UsageBar from items/. Behavior-preserving; all gates
green (lint, typecheck, build, 898 tests — 880 prior + 18 new).
